third_party/WebKit/LayoutTests/http/tests/security/contentTypeOptions/block-video-as-script.html

<!DOCTYPE html>
<head>
    <title>Block 'video/*' when served as script.</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
      window.scriptsSuccessfullyLoaded = 0;
    </script>
</head>
<body>
    <script src="resources/script-with-header.pl?mime=video/avi&amp;options=invalid"></script>
    <script src="resources/script-with-header.pl?mime=ViDeO/fli&amp;options=invalid"></script>
    <script src="resources/script-with-header.pl?mime=video/x-motion-jpeg&amp;options=invalid"></script>
    <script src="resources/script-with-header.pl?mime=video/whatever+xml&amp;options=invalid"></script>
    <script>
      test(function () {
        assert_equals(window.scriptsSuccessfullyLoaded, 0);
      }, "All video/* scripts should be blocked.");
    </script>
</body>
</html>
